refactor: split theme type augmentation per site (#1620)
