+++ packages/shared/@types/theme.d.ts
@@ -1,0 +1,11 @@
+export {};
+
+/**
+ * Chip prop options shared by every site.
+ */
+declare module "@mui/material/Chip" {
+  interface ChipPropsColorOverrides {
+    alert: true;
+    none: true;
+  }
+}
